Get android presence viewer working
