+++ dittopresenceviewer/build.gradle
@@ -65,7 +65,7 @@ dependencies {
     implementation "androidx.compose.material:material-icons-extended:$compose_ui_version"
 //    implementation "androidx.compose.runtime:runtime-livedata:$compose_ui_version"
 //    implementation "androidx.compose.webview:webview:$compose_ui_version"
-
+    implementation "androidx.webkit:webkit:1.7.0"
 
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
